Docs: cleanup and linking!
This iterates on docs a bit. Once noticing <remarks> doesn't render after the first one, I realized we were dropping ref links...and it'd be better if they were clickable. This globally makes them clickable from intellisense hovers.
