+++ src/StackExchange.Redis/Interfaces/IDatabase.cs
@@ -41,7 +41,7 @@ namespace StackExchange.Redis
         /// <param name="timeoutMilliseconds">The timeout to use for the transfer.</param>
         /// <param name="migrateOptions">The options to use for this migration.</param>
         /// <param name="flags">The flags to use for this operation.</param>
-        /// <remarks>https://redis.io/commands/MIGRATE</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/migrate"/></remarks>
         void KeyMigrate(RedisKey key, EndPoint toServer, int toDatabase = 0, int timeoutMilliseconds = 0, MigrateOptions migrateOptions = MigrateOptions.None, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -51,7 +51,7 @@ namespace StackExchange.Redis
         /// <param name="key">The key to debug.</param>
         /// <param name="flags">The flags to use for this migration.</param>
         /// <returns>The raw output from DEBUG OBJECT.</returns>
-        /// <remarks>https://redis.io/commands/debug-object</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/debug-object"/></remarks>
         RedisValue DebugObject(RedisKey key, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -65,7 +65,7 @@ namespace StackExchange.Redis
         /// <param name="member">The value to set at this entry.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns><see langword="true"/> if the specified member was not already present in the set, else <see langword="false"/>.</returns>
-        /// <remarks>https://redis.io/commands/geoadd</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/geoadd"/></remarks>
         bool GeoAdd(RedisKey key, double longitude, double latitude, RedisValue member, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -77,7 +77,7 @@ namespace StackExchange.Redis
         /// <param name="value">The geo value to store.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns><see langword="true"/> if the specified member was not already present in the set, else <see langword="false"/>.</returns>
-        /// <remarks>https://redis.io/commands/geoadd</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/geoadd"/></remarks>
         bool GeoAdd(RedisKey key, GeoEntry value, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -89,7 +89,7 @@ namespace StackExchange.Redis
         /// <param name="values">The geo values add to the set.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The number of elements that were added to the set, not including all the elements already present into the set.</returns>
-        /// <remarks>https://redis.io/commands/geoadd</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/geoadd"/></remarks>
         long GeoAdd(RedisKey key, GeoEntry[] values, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -100,7 +100,7 @@ namespace StackExchange.Redis
         /// <param name="member">The geo value to remove.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns><see langword="true"/> if the member existed in the sorted set and was removed, else <see langword="false"/>.</returns>
-        /// <remarks>https://redis.io/commands/zrem</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/zrem"/></remarks>
         bool GeoRemove(RedisKey key, RedisValue member, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -112,7 +112,7 @@ namespace StackExchange.Redis
         /// <param name="unit">The unit of distance to return (defaults to meters).</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The command returns the distance as a double (represented as a string) in the specified unit, or <see langword="null"/> if one or both the elements are missing.</returns>
-        /// <remarks>https://redis.io/commands/geodist</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/geodist"/></remarks>
         double? GeoDistance(RedisKey key, RedisValue member1, RedisValue member2, GeoUnit unit = GeoUnit.Meters, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -122,7 +122,7 @@ namespace StackExchange.Redis
         /// <param name="members">The members to get.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The command returns an array where each element is the Geohash corresponding to each member name passed as argument to the command.</returns>
-        /// <remarks>https://redis.io/commands/geohash</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/geohash"/></remarks>
         string?[] GeoHash(RedisKey key, RedisValue[] members, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -132,7 +132,7 @@ namespace StackExchange.Redis
         /// <param name="member">The member to get.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The command returns an array where each element is the Geohash corresponding to each member name passed as argument to the command.</returns>
-        /// <remarks>https://redis.io/commands/geohash</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/geohash"/></remarks>
         string? GeoHash(RedisKey key, RedisValue member, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -145,7 +145,7 @@ namespace StackExchange.Redis
         /// The command returns an array where each element is a two elements array representing longitude and latitude (x,y) of each member name passed as argument to the command.
         /// Non existing elements are reported as NULL elements of the array.
         /// </returns>
-        /// <remarks>https://redis.io/commands/geopos</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/geopos"/></remarks>
         GeoPosition?[] GeoPosition(RedisKey key, RedisValue[] members, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -158,7 +158,7 @@ namespace StackExchange.Redis
         /// The command returns an array where each element is a two elements array representing longitude and latitude (x,y) of each member name passed as argument to the command.
         /// Non existing elements are reported as NULL elements of the array.
         /// </returns>
-        /// <remarks>https://redis.io/commands/geopos</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/geopos"/></remarks>
         GeoPosition? GeoPosition(RedisKey key, RedisValue member, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -174,7 +174,7 @@ namespace StackExchange.Redis
         /// <param name="options">The search options to use.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The results found within the radius, if any.</returns>
-        /// <remarks>https://redis.io/commands/georadius</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/georadius"/></remarks>
         GeoRadiusResult[] GeoRadius(RedisKey key, RedisValue member, double radius, GeoUnit unit = GeoUnit.Meters, int count = -1, Order? order = null, GeoRadiusOptions options = GeoRadiusOptions.Default, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -191,7 +191,7 @@ namespace StackExchange.Redis
         /// <param name="options">The search options to use.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The results found within the radius, if any.</returns>
-        /// <remarks>https://redis.io/commands/georadius</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/georadius"/></remarks>
         GeoRadiusResult[] GeoRadius(RedisKey key, double longitude, double latitude, double radius, GeoUnit unit = GeoUnit.Meters, int count = -1, Order? order = null, GeoRadiusOptions options = GeoRadiusOptions.Default, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -204,10 +204,10 @@ namespace StackExchange.Redis
         /// <param name="count">The maximum number of results to pull back.</param>
         /// <param name="demandClosest">Whether or not to terminate the search after finding <paramref name="count"/> results. Must be true of count is -1.</param>
         /// <param name="order">The order to sort by (defaults to unordered).</param>
-        /// <param name="options">The search options to use</param>
+        /// <param name="options">The search options to use.</param>
         /// <param name="flags">The flags for this operation.</param>
         /// <returns>The results found within the shape, if any.</returns>
-        /// <remarks>https://redis.io/commands/geosearch</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/geosearch"/></remarks>
         GeoRadiusResult[] GeoSearch(RedisKey key, RedisValue member, GeoSearchShape shape, int count = -1, bool demandClosest = true, Order? order = null, GeoRadiusOptions options = GeoRadiusOptions.Default, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -221,10 +221,10 @@ namespace StackExchange.Redis
         /// <param name="count">The maximum number of results to pull back.</param>
         /// <param name="demandClosest">Whether or not to terminate the search after finding <paramref name="count"/> results. Must be true of count is -1.</param>
         /// <param name="order">The order to sort by (defaults to unordered).</param>
-        /// <param name="options">The search options to use</param>
+        /// <param name="options">The search options to use.</param>
         /// <param name="flags">The flags for this operation.</param>
         /// <returns>The results found within the shape, if any.</returns>
-        /// /// <remarks>https://redis.io/commands/geosearch</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/geosearch"/></remarks>
         GeoRadiusResult[] GeoSearch(RedisKey key, double longitude, double latitude, GeoSearchShape shape, int count = -1, bool demandClosest = true, Order? order = null, GeoRadiusOptions options = GeoRadiusOptions.Default, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -241,7 +241,7 @@ namespace StackExchange.Redis
         /// <param name="storeDistances">If set to true, the resulting set will be a regular sorted-set containing only distances, rather than a geo-encoded sorted-set.</param>
         /// <param name="flags">The flags for this operation.</param>
         /// <returns>The size of the set stored at <paramref name="destinationKey"/>.</returns>
-        /// <remarks>https://redis.io/commands/geosearchstore</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/geosearchstore"/></remarks>
         long GeoSearchAndStore(RedisKey sourceKey, RedisKey destinationKey, RedisValue member, GeoSearchShape shape, int count = -1, bool demandClosest = true, Order? order = null, bool storeDistances = false, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -259,7 +259,7 @@ namespace StackExchange.Redis
         /// <param name="storeDistances">If set to true, the resulting set will be a regular sorted-set containing only distances, rather than a geo-encoded sorted-set.</param>
         /// <param name="flags">The flags for this operation.</param>
         /// <returns>The size of the set stored at <paramref name="destinationKey"/>.</returns>
-        /// <remarks>https://redis.io/commands/geosearchstore</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/geosearchstore"/></remarks>
         long GeoSearchAndStore(RedisKey sourceKey, RedisKey destinationKey, double longitude, double latitude, GeoSearchShape shape, int count = -1, bool demandClosest = true, Order? order = null, bool storeDistances = false, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -272,8 +272,10 @@ namespace StackExchange.Redis
         /// <param name="value">The amount to decrement by.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The value at field after the decrement operation.</returns>
-        /// <remarks>The range of values supported by HINCRBY is limited to 64 bit signed integers.</remarks>
-        /// <remarks>https://redis.io/commands/hincrby</remarks>
+        /// <remarks>
+        /// <para>The range of values supported by HINCRBY is limited to 64 bit signed integers.</para>
+        /// <para><seealso href="https://redis.io/commands/hincrby"/></para>
+        /// </remarks>
         long HashDecrement(RedisKey key, RedisValue hashField, long value = 1, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -285,8 +287,10 @@ namespace StackExchange.Redis
         /// <param name="value">The amount to decrement by.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The value at field after the decrement operation.</returns>
-        /// <remarks>The precision of the output is fixed at 17 digits after the decimal point regardless of the actual internal precision of the computation.</remarks>
-        /// <remarks>https://redis.io/commands/hincrbyfloat</remarks>
+        /// <remarks>
+        /// <para>The precision of the output is fixed at 17 digits after the decimal point regardless of the actual internal precision of the computation.</para>
+        /// <para><seealso href="https://redis.io/commands/hincrbyfloat"/></para>
+        /// </remarks>
         double HashDecrement(RedisKey key, RedisValue hashField, double value, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -297,7 +301,7 @@ namespace StackExchange.Redis
         /// <param name="hashField">The field in the hash to delete.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The number of fields that were removed.</returns>
-        /// <remarks>https://redis.io/commands/hdel</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/hdel"/></remarks>
         bool HashDelete(RedisKey key, RedisValue hashField, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -308,7 +312,7 @@ namespace StackExchange.Redis
         /// <param name="hashFields">The fields in the hash to delete.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The number of fields that were removed.</returns>
-        /// <remarks>https://redis.io/commands/hdel</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/hdel"/></remarks>
         long HashDelete(RedisKey key, RedisValue[] hashFields, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -318,7 +322,7 @@ namespace StackExchange.Redis
         /// <param name="hashField">The field in the hash to check.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns><see langword="true"/> if the hash contains field, <see langword="false"/> if the hash does not contain field, or key does not exist.</returns>
-        /// <remarks>https://redis.io/commands/hexists</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/hexists"/></remarks>
         bool HashExists(RedisKey key, RedisValue hashField, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -328,7 +332,7 @@ namespace StackExchange.Redis
         /// <param name="hashField">The field in the hash to get.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The value associated with field, or nil when field is not present in the hash or key does not exist.</returns>
-        /// <remarks>https://redis.io/commands/hget</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/hget"/></remarks>
         RedisValue HashGet(RedisKey key, RedisValue hashField, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -338,7 +342,7 @@ namespace StackExchange.Redis
         /// <param name="hashField">The field in the hash to get.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The value associated with field, or nil when field is not present in the hash or key does not exist.</returns>
-        /// <remarks>https://redis.io/commands/hget</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/hget"/></remarks>
         Lease<byte>? HashGetLease(RedisKey key, RedisValue hashField, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -349,7 +353,7 @@ namespace StackExchange.Redis
         /// <param name="hashFields">The fields in the hash to get.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>List of values associated with the given fields, in the same order as they are requested.</returns>
-        /// <remarks>https://redis.io/commands/hmget</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/hmget"/></remarks>
         RedisValue[] HashGet(RedisKey key, RedisValue[] hashFields, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -358,7 +362,7 @@ namespace StackExchange.Redis
         /// <param name="key">The key of the hash to get all entries from.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>List of fields and their values stored in the hash, or an empty list when key does not exist.</returns>
-        /// <remarks>https://redis.io/commands/hgetall</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/hgetall"/></remarks>
         HashEntry[] HashGetAll(RedisKey key, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -371,8 +375,10 @@ namespace StackExchange.Redis
         /// <param name="value">The amount to increment by.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The value at field after the increment operation.</returns>
-        /// <remarks>The range of values supported by HINCRBY is limited to 64 bit signed integers.</remarks>
-        /// <remarks>https://redis.io/commands/hincrby</remarks>
+        /// <remarks>
+        /// <para>The range of values supported by <c>HINCRBY</c> is limited to 64 bit signed integers.</para>
+        /// <para><seealso href="https://redis.io/commands/hincrby"/></para>
+        /// </remarks>
         long HashIncrement(RedisKey key, RedisValue hashField, long value = 1, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -384,8 +390,10 @@ namespace StackExchange.Redis
         /// <param name="value">The amount to increment by.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The value at field after the increment operation.</returns>
-        /// <remarks>The precision of the output is fixed at 17 digits after the decimal point regardless of the actual internal precision of the computation.</remarks>
-        /// <remarks>https://redis.io/commands/hincrbyfloat</remarks>
+        /// <remarks>
+        /// <para>The precision of the output is fixed at 17 digits after the decimal point regardless of the actual internal precision of the computation.</para>
+        /// <para><seealso href="https://redis.io/commands/hincrbyfloat"/></para>
+        /// </remarks>
         double HashIncrement(RedisKey key, RedisValue hashField, double value, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -394,7 +402,7 @@ namespace StackExchange.Redis
         /// <param name="key">The key of the hash.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>List of fields in the hash, or an empty list when key does not exist.</returns>
-        /// <remarks>https://redis.io/commands/hkeys</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/hkeys"/></remarks>
         RedisValue[] HashKeys(RedisKey key, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -403,7 +411,7 @@ namespace StackExchange.Redis
         /// <param name="key">The key of the hash.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The number of fields in the hash, or 0 when key does not exist.</returns>
-        /// <remarks>https://redis.io/commands/hlen</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/hlen"/></remarks>
         long HashLength(RedisKey key, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -412,7 +420,7 @@ namespace StackExchange.Redis
         /// <param name="key">The key of the hash.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>A random hash field name or <see cref="RedisValue.Null"/> if the hash does not exist.</returns>
-        /// <remarks>https://redis.io/commands/hrandfield</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/hrandfield"/></remarks>
         RedisValue HashRandomField(RedisKey key, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -422,7 +430,7 @@ namespace StackExchange.Redis
         /// <param name="count">The number of fields to return.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>An array of hash field names of size of at most <paramref name="count"/>, or <see cref="Array.Empty{RedisValue}"/> if the hash does not exist.</returns>
-        /// <remarks>https://redis.io/commands/hrandfield</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/hrandfield"/></remarks>
         RedisValue[] HashRandomFields(RedisKey key, long count, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -432,7 +440,7 @@ namespace StackExchange.Redis
         /// <param name="count">The number of fields to return.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>An array of hash entries of size of at most <paramref name="count"/>, or <see cref="Array.Empty{HashEntry}"/> if the hash does not exist.</returns>
-        /// <remarks>https://redis.io/commands/hrandfield</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/hrandfield"/></remarks>
         HashEntry[] HashRandomFieldsWithValues(RedisKey key, long count, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -443,7 +451,7 @@ namespace StackExchange.Redis
         /// <param name="pageSize">The page size to iterate by.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>Yields all elements of the hash matching the pattern.</returns>
-        /// <remarks>https://redis.io/commands/hscan</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/hscan"/></remarks>
         IEnumerable<HashEntry> HashScan(RedisKey key, RedisValue pattern, int pageSize, CommandFlags flags);
 
         /// <summary>
@@ -457,7 +465,7 @@ namespace StackExchange.Redis
         /// <param name="pageOffset">The page offset to start at.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>Yields all elements of the hash matching the pattern.</returns>
-        /// <remarks>https://redis.io/commands/hscan</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/hscan"/></remarks>
         IEnumerable<HashEntry> HashScan(RedisKey key, RedisValue pattern = default, int pageSize = RedisBase.CursorUtils.DefaultLibraryPageSize, long cursor = RedisBase.CursorUtils.Origin, int pageOffset = 0, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -468,7 +476,7 @@ namespace StackExchange.Redis
         /// <param name="key">The key of the hash.</param>
         /// <param name="hashFields">The entries to set in the hash.</param>
         /// <param name="flags">The flags to use for this operation.</param>
-        /// <remarks>https://redis.io/commands/hmset</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/hmset"/></remarks>
         void HashSet(RedisKey key, HashEntry[] hashFields, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -482,8 +490,10 @@ namespace StackExchange.Redis
         /// <param name="when">Which conditions under which to set the field value (defaults to always).</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns><see langword="true"/> if field is a new field in the hash and value was set, <see langword="false"/> if field already exists in the hash and the value was updated.</returns>
-        /// <remarks>https://redis.io/commands/hset</remarks>
-        /// <remarks>https://redis.io/commands/hsetnx</remarks>
+        /// <remarks>
+        /// <seealso href="https://redis.io/commands/hset"/>,
+        /// <seealso href="https://redis.io/commands/hsetnx"/>
+        /// </remarks>
         bool HashSet(RedisKey key, RedisValue hashField, RedisValue value, When when = When.Always, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -493,7 +503,7 @@ namespace StackExchange.Redis
         /// <param name="hashField">The field containing the string</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The length of the string at field, or 0 when key does not exist.</returns>
-        /// <remarks>https://redis.io/commands/hstrlen</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/hstrlen"/></remarks>
         long HashStringLength(RedisKey key, RedisValue hashField, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -502,7 +512,7 @@ namespace StackExchange.Redis
         /// <param name="key">The key of the hash.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>List of values in the hash, or an empty list when key does not exist.</returns>
-        /// <remarks>https://redis.io/commands/hvals</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/hvals"/></remarks>
         RedisValue[] HashValues(RedisKey key, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -512,7 +522,7 @@ namespace StackExchange.Redis
         /// <param name="value">The value to add.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns><see langword="true"/> if at least 1 HyperLogLog internal register was altered, <see langword="false"/> otherwise.</returns>
-        /// <remarks>https://redis.io/commands/pfadd</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/pfadd"/></remarks>
         bool HyperLogLogAdd(RedisKey key, RedisValue value, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -522,7 +532,7 @@ namespace StackExchange.Redis
         /// <param name="values">The values to add.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns><see langword="true"/> if at least 1 HyperLogLog internal register was altered, <see langword="false"/> otherwise.</returns>
-        /// <remarks>https://redis.io/commands/pfadd</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/pfadd"/></remarks>
         bool HyperLogLogAdd(RedisKey key, RedisValue[] values, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -531,7 +541,7 @@ namespace StackExchange.Redis
         /// <param name="key">The key of the hyperloglog.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The approximated number of unique elements observed via HyperLogLogAdd.</returns>
-        /// <remarks>https://redis.io/commands/pfcount</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/pfcount"/></remarks>
         long HyperLogLogLength(RedisKey key, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -540,7 +550,7 @@ namespace StackExchange.Redis
         /// <param name="keys">The keys of the hyperloglogs.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The approximated number of unique elements observed via HyperLogLogAdd.</returns>
-        /// <remarks>https://redis.io/commands/pfcount</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/pfcount"/></remarks>
         long HyperLogLogLength(RedisKey[] keys, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -550,7 +560,7 @@ namespace StackExchange.Redis
         /// <param name="first">The key of the first hyperloglog to merge.</param>
         /// <param name="second">The key of the first hyperloglog to merge.</param>
         /// <param name="flags">The flags to use for this operation.</param>
-        /// <remarks>https://redis.io/commands/pfmerge</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/pfmerge"/></remarks>
         void HyperLogLogMerge(RedisKey destination, RedisKey first, RedisKey second, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -559,7 +569,7 @@ namespace StackExchange.Redis
         /// <param name="destination">The key of the merged hyperloglog.</param>
         /// <param name="sourceKeys">The keys of the hyperloglogs to merge.</param>
         /// <param name="flags">The flags to use for this operation.</param>
-        /// <remarks>https://redis.io/commands/pfmerge</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/pfmerge"/></remarks>
         void HyperLogLogMerge(RedisKey destination, RedisKey[] sourceKeys, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -579,7 +589,7 @@ namespace StackExchange.Redis
         /// <param name="replace">Whether to overwrite an existing values at <paramref name="destinationKey"/>. If <see langword="false"/> and the key exists, the copy will not succeed.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns><see langword="true"/> if key was copied. <see langword="false"/> if key was not copied.</returns>
-        /// <remarks>https://redis.io/commands/copy</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/copy"/></remarks>
         bool KeyCopy(RedisKey sourceKey, RedisKey destinationKey, int destinationDatabase = -1, bool replace = false, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -589,8 +599,10 @@ namespace StackExchange.Redis
         /// <param name="key">The key to delete.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns><see langword="true"/> if the key was removed.</returns>
-        /// <remarks>https://redis.io/commands/del</remarks>
-        /// <remarks>https://redis.io/commands/unlink</remarks>
+        /// <remarks>
+        /// <seealso href="https://redis.io/commands/del"/>,
+        /// <seealso href="https://redis.io/commands/unlink"/>
+        /// </remarks>
         bool KeyDelete(RedisKey key, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -600,8 +612,10 @@ namespace StackExchange.Redis
         /// <param name="keys">The keys to delete.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The number of keys that were removed.</returns>
-        /// <remarks>https://redis.io/commands/del</remarks>
-        /// <remarks>https://redis.io/commands/unlink</remarks>
+        /// <remarks>
+        /// <seealso href="https://redis.io/commands/del"/>,
+        /// <seealso href="https://redis.io/commands/unlink"/>
+        /// </remarks>
         long KeyDelete(RedisKey[] keys, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -611,7 +625,7 @@ namespace StackExchange.Redis
         /// <param name="key">The key to dump.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The serialized value.</returns>
-        /// <remarks>https://redis.io/commands/dump</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/dump"/></remarks>
         byte[]? KeyDump(RedisKey key, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -620,7 +634,7 @@ namespace StackExchange.Redis
         /// <param name="key">The key to dump.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The Redis encoding for the value or <see langword="null"/> is the key does not exist.</returns>
-        /// <remarks>https://redis.io/commands/object-encoding</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/object-encoding"/></remarks>
         string? KeyEncoding(RedisKey key, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -629,7 +643,7 @@ namespace StackExchange.Redis
         /// <param name="key">The key to check.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns><see langword="true"/> if the key exists. <see langword="false"/> if the key does not exist.</returns>
-        /// <remarks>https://redis.io/commands/exists</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/exists"/></remarks>
         bool KeyExists(RedisKey key, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -638,7 +652,7 @@ namespace StackExchange.Redis
         /// <param name="keys">The keys to check.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The number of keys that existed.</returns>
-        /// <remarks>https://redis.io/commands/exists</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/exists"/></remarks>
         long KeyExists(RedisKey[] keys, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -658,12 +672,15 @@ namespace StackExchange.Redis
         /// </para>
         /// <para>
         /// Since Redis 2.1.3, you can update the timeout of a key.
-        /// It is also possible to remove the timeout using the PERSIST command. See the page on key expiry for more information.
+        /// It is also possible to remove the timeout using the PERSIST command.
+        /// See the page on key expiry for more information.
+        /// </para>
+        /// <para>
+        /// <seealso href="https://redis.io/commands/expire"/>,
+        /// <seealso href="https://redis.io/commands/pexpire"/>,
+        /// <seealso href="https://redis.io/commands/persist"/>
         /// </para>
         /// </remarks>
-        /// <remarks>https://redis.io/commands/expire</remarks>
-        /// <remarks>https://redis.io/commands/pexpire</remarks>
-        /// <remarks>https://redis.io/commands/persist</remarks>
         bool KeyExpire(RedisKey key, TimeSpan? expiry, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -676,8 +693,10 @@ namespace StackExchange.Redis
         /// <param name="when">Since Redis 7.0.0, you can choose under which condition the expiration will be set using <see cref="ExpireWhen"/>.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns><see langword="true"/> if the timeout was set. <see langword="false"/> if key does not exist or the timeout could not be set.</returns>
-        /// <remarks>https://redis.io/commands/expire</remarks>
-        /// <remarks>https://redis.io/commands/pexpire</remarks>
+        /// <remarks>
+        /// <seealso href="https://redis.io/commands/expire"/>,
+        /// <seealso href="https://redis.io/commands/pexpire"/>
+        /// </remarks>
         bool KeyExpire(RedisKey key, TimeSpan? expiry, ExpireWhen when, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -697,12 +716,15 @@ namespace StackExchange.Redis
         /// </para>
         /// <para>
         /// Since Redis 2.1.3, you can update the timeout of a key.
-        /// It is also possible to remove the timeout using the PERSIST command. See the page on key expiry for more information.
+        /// It is also possible to remove the timeout using the PERSIST command.
+        /// See the page on key expiry for more information.
+        /// </para>
+        /// <para>
+        /// <seealso href="https://redis.io/commands/expireat"/>,
+        /// <seealso href="https://redis.io/commands/pexpireat"/>,
+        /// <seealso href="https://redis.io/commands/persist"/>
         /// </para>
         /// </remarks>
-        /// <remarks>https://redis.io/commands/expireat</remarks>
-        /// <remarks>https://redis.io/commands/pexpireat</remarks>
-        /// <remarks>https://redis.io/commands/persist</remarks>
         bool KeyExpire(RedisKey key, DateTime? expiry, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -715,8 +737,10 @@ namespace StackExchange.Redis
         /// <param name="when">Since Redis 7.0.0, you can choose under which condition the expiration will be set using <see cref="ExpireWhen"/>.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns><see langword="true"/> if the timeout was set. <see langword="false"/> if key does not exist or the timeout could not be set.</returns>
-        /// <remarks>https://redis.io/commands/expire</remarks>
-        /// <remarks>https://redis.io/commands/pexpire</remarks>
+        /// <remarks>
+        /// <seealso href="https://redis.io/commands/expire"/>,
+        /// <seealso href="https://redis.io/commands/pexpire"/>
+        /// </remarks>
         bool KeyExpire(RedisKey key, DateTime? expiry, ExpireWhen when, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -725,8 +749,10 @@ namespace StackExchange.Redis
         /// <param name="key">The key to get the expiration for.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The time at which the given key will expire, or <see langword="null"/> if the key does not exist or has no associated expiration time.</returns>
-        /// <remarks>https://redis.io/commands/expiretime</remarks>
-        /// <remarks>https://redis.io/commands/pexpiretime</remarks>
+        /// <remarks>
+        /// <seealso href="https://redis.io/commands/expiretime"/>,
+        /// <seealso href="https://redis.io/commands/pexpiretime"/>
+        /// </remarks>
         DateTime? KeyExpireTime(RedisKey key, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -735,7 +761,7 @@ namespace StackExchange.Redis
         /// <param name="key">The key to get the time of.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The time since the object stored at the specified key is idle.</returns>
-        /// <remarks>https://redis.io/commands/object</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/object"/></remarks>
         TimeSpan? KeyIdleTime(RedisKey key, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -747,7 +773,7 @@ namespace StackExchange.Redis
         /// <param name="database">The database to move the key to.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns><see langword="true"/> if key was moved. <see langword="false"/> if key was not moved.</returns>
-        /// <remarks>https://redis.io/commands/move</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/move"/></remarks>
         bool KeyMove(RedisKey key, int database, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -756,7 +782,7 @@ namespace StackExchange.Redis
         /// <param name="key">The key to persist.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns><see langword="true"/> if the timeout was removed. <see langword="false"/> if key does not exist or does not have an associated timeout.</returns>
-        /// <remarks>https://redis.io/commands/persist</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/persist"/></remarks>
         bool KeyPersist(RedisKey key, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -764,7 +790,7 @@ namespace StackExchange.Redis
         /// </summary>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The random key, or nil when the database is empty.</returns>
-        /// <remarks>https://redis.io/commands/randomkey</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/randomkey"/></remarks>
         RedisKey KeyRandom(CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -773,7 +799,7 @@ namespace StackExchange.Redis
         /// <param name="key">The key to get a reference count for.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The number of references (<see langword="Null"/> if the key does not exist).</returns>
-        /// <remarks>https://redis.io/commands/object-refcount</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/object-refcount"/></remarks>
         long? KeyRefCount(RedisKey key, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -785,8 +811,10 @@ namespace StackExchange.Redis
         /// <param name="when">What conditions to rename under (defaults to always).</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns><see langword="true"/> if the key was renamed, <see langword="false"/> otherwise.</returns>
-        /// <remarks>https://redis.io/commands/rename</remarks>
-        /// <remarks>https://redis.io/commands/renamenx</remarks>
+        /// <remarks>
+        /// <seealso href="https://redis.io/commands/rename"/>,
+        /// <seealso href="https://redis.io/commands/renamenx"/>
+        /// </remarks>
         bool KeyRename(RedisKey key, RedisKey newKey, When when = When.Always, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -797,7 +825,7 @@ namespace StackExchange.Redis
         /// <param name="value">The value of the key.</param>
         /// <param name="expiry">The expiry to set.</param>
         /// <param name="flags">The flags to use for this operation.</param>
-        /// <remarks>https://redis.io/commands/restore</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/restore"/></remarks>
         void KeyRestore(RedisKey key, byte[] value, TimeSpan? expiry = null, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -807,7 +835,7 @@ namespace StackExchange.Redis
         /// <param name="key">The key to check.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>TTL, or nil when key does not exist or does not have a timeout.</returns>
-        /// <remarks>https://redis.io/commands/ttl</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/ttl"/></remarks>
         TimeSpan? KeyTimeToLive(RedisKey key, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -816,7 +844,7 @@ namespace StackExchange.Redis
         /// <param name="key">The key to touch.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns><see langword="true"/> if the key was touched, <see langword="false"/> otherwise.</returns>
-        /// <remarks>https://redis.io/commands/touch</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/touch"/></remarks>
         bool KeyTouch(RedisKey key, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -825,7 +853,7 @@ namespace StackExchange.Redis
         /// <param name="keys">The keys to touch.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The number of keys that were touched.</returns>
-        /// <remarks>https://redis.io/commands/touch</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/touch"/></remarks>
         long KeyTouch(RedisKey[] keys, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -835,7 +863,7 @@ namespace StackExchange.Redis
         /// <param name="key">The key to get the type of.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>Type of key, or none when key does not exist.</returns>
-        /// <remarks>https://redis.io/commands/type</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/type"/></remarks>
         RedisType KeyType(RedisKey key, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -848,7 +876,7 @@ namespace StackExchange.Redis
         /// <param name="index">The index position to get the value at.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The requested element, or nil when index is out of range.</returns>
-        /// <remarks>https://redis.io/commands/lindex</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/lindex"/></remarks>
         RedisValue ListGetByIndex(RedisKey key, long index, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -860,7 +888,7 @@ namespace StackExchange.Redis
         /// <param name="value">The value to insert.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The length of the list after the insert operation, or -1 when the value pivot was not found.</returns>
-        /// <remarks>https://redis.io/commands/linsert</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/linsert"/></remarks>
         long ListInsertAfter(RedisKey key, RedisValue pivot, RedisValue value, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -872,7 +900,7 @@ namespace StackExchange.Redis
         /// <param name="value">The value to insert.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The length of the list after the insert operation, or -1 when the value pivot was not found.</returns>
-        /// <remarks>https://redis.io/commands/linsert</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/linsert"/></remarks>
         long ListInsertBefore(RedisKey key, RedisValue pivot, RedisValue value, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -881,7 +909,7 @@ namespace StackExchange.Redis
         /// <param name="key">The key of the list.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The value of the first element, or nil when key does not exist.</returns>
-        /// <remarks>https://redis.io/commands/lpop</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/lpop"/></remarks>
         RedisValue ListLeftPop(RedisKey key, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -892,7 +920,7 @@ namespace StackExchange.Redis
         /// <param name="count">The number of elements to remove</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>Array of values that were popped, or nil if the key doesn't exist.</returns>
-        /// <remarks>https://redis.io/commands/lpop</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/lpop"/></remarks>
         RedisValue[] ListLeftPop(RedisKey key, long count, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -905,6 +933,7 @@ namespace StackExchange.Redis
         /// <param name="maxLength">The maximum number of elements to scan through before stopping, defaults to 0 (a full scan of the list.)</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The 0-based index of the first matching element, or -1 if not found.</returns>
+        /// <remarks><seealso href="https://redis.io/commands/lpos"/></remarks>
         long ListPosition(RedisKey key, RedisValue element, long rank = 1, long maxLength = 0, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -918,6 +947,7 @@ namespace StackExchange.Redis
         /// <param name="maxLength">The maximum number of elements to scan through before stopping, defaults to 0 (a full scan of the list.)</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>An array of at most <paramref name="count"/> of indexes of matching elements. If none are found, and empty array is returned.</returns>
+        /// <remarks><seealso href="https://redis.io/commands/lpos"/></remarks>
         long[] ListPositions(RedisKey key, RedisValue element, long count, long rank = 1, long maxLength = 0, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -929,8 +959,10 @@ namespace StackExchange.Redis
         /// <param name="when">Which conditions to add to the list under (defaults to always).</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The length of the list after the push operations.</returns>
-        /// <remarks>https://redis.io/commands/lpush</remarks>
-        /// <remarks>https://redis.io/commands/lpushx</remarks>
+        /// <remarks>
+        /// <seealso href="https://redis.io/commands/lpush"/>,
+        /// <seealso href="https://redis.io/commands/lpushx"/>
+        /// </remarks>
         long ListLeftPush(RedisKey key, RedisValue value, When when = When.Always, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -942,21 +974,23 @@ namespace StackExchange.Redis
         /// <param name="when">Which conditions to add to the list under (defaults to always).</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The length of the list after the push operations.</returns>
-        /// <remarks>https://redis.io/commands/lpush</remarks>
-        /// <remarks>https://redis.io/commands/lpushx</remarks>
+        /// <remarks>
+        /// <seealso href="https://redis.io/commands/lpush"/>,
+        /// <seealso href="https://redis.io/commands/lpushx"/>
+        /// </remarks>
         long ListLeftPush(RedisKey key, RedisValue[] values, When when = When.Always, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
         /// Insert all the specified values at the head of the list stored at key.
         /// If key does not exist, it is created as empty list before performing the push operations.
         /// Elements are inserted one after the other to the head of the list, from the leftmost element to the rightmost element.
-        /// So for instance the command LPUSH mylist a b c will result into a list containing c as first element, b as second element and a as third element.
+        /// So for instance the command <c>LPUSH mylist a b c</c> will result into a list containing c as first element, b as second element and a as third element.
         /// </summary>
         /// <param name="key">The key of the list.</param>
         /// <param name="values">The values to add to the head of the list.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The length of the list after the push operations.</returns>
-        /// <remarks>https://redis.io/commands/lpush</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/lpush"/></remarks>
         long ListLeftPush(RedisKey key, RedisValue[] values, CommandFlags flags);
 
         /// <summary>
@@ -965,7 +999,7 @@ namespace StackExchange.Redis
         /// <param name="key">The key of the list.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The length of the list at key.</returns>
-        /// <remarks>https://redis.io/commands/llen</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/llen"/></remarks>
         long ListLength(RedisKey key, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -978,7 +1012,7 @@ namespace StackExchange.Redis
         /// <param name="destinationSide">What side of the <paramref name="destinationKey"/> list to move to.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The element being popped and pushed or <see cref="RedisValue.Null"/> if there is no element to move.</returns>
-        /// <remarks>https://redis.io/commands/lmove</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/lmove"/></remarks>
         RedisValue ListMove(RedisKey sourceKey, RedisKey destinationKey, ListSide sourceSide, ListSide destinationSide, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -992,7 +1026,7 @@ namespace StackExchange.Redis
         /// <param name="stop">The stop index of the list.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>List of elements in the specified range.</returns>
-        /// <remarks>https://redis.io/commands/lrange</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/lrange"/></remarks>
         RedisValue[] ListRange(RedisKey key, long start = 0, long stop = -1, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -1009,7 +1043,7 @@ namespace StackExchange.Redis
         /// <param name="count">The count behavior (see method summary).</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The number of removed elements.</returns>
-        /// <remarks>https://redis.io/commands/lrem</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/lrem"/></remarks>
         long ListRemove(RedisKey key, RedisValue value, long count = 0, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -1018,7 +1052,7 @@ namespace StackExchange.Redis
         /// <param name="key">The key of the list.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The element being popped.</returns>
-        /// <remarks>https://redis.io/commands/rpop</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/rpop"/></remarks>
         RedisValue ListRightPop(RedisKey key, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -1029,7 +1063,7 @@ namespace StackExchange.Redis
         /// <param name="count">The number of elements to pop</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>Array of values that were popped, or nil if the key doesn't exist.</returns>
-        /// <remarks>https://redis.io/commands/rpop</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/rpop"/></remarks>
         RedisValue[] ListRightPop(RedisKey key, long count, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -1039,7 +1073,7 @@ namespace StackExchange.Redis
         /// <param name="destination">The key of the destination list.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The element being popped and pushed.</returns>
-        /// <remarks>https://redis.io/commands/rpoplpush</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/rpoplpush"/></remarks>
         RedisValue ListRightPopLeftPush(RedisKey source, RedisKey destination, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -1051,8 +1085,10 @@ namespace StackExchange.Redis
         /// <param name="when">Which conditions to add to the list under.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The length of the list after the push operation.</returns>
-        /// <remarks>https://redis.io/commands/rpush</remarks>
-        /// <remarks>https://redis.io/commands/rpushx</remarks>
+        /// <remarks>
+        /// <seealso href="https://redis.io/commands/rpush"/>,
+        /// <seealso href="https://redis.io/commands/rpushx"/>
+        /// </remarks>
         long ListRightPush(RedisKey key, RedisValue value, When when = When.Always, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -1064,21 +1100,23 @@ namespace StackExchange.Redis
         /// <param name="when">Which conditions to add to the list under.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The length of the list after the push operation.</returns>
-        /// <remarks>https://redis.io/commands/rpush</remarks>
-        /// <remarks>https://redis.io/commands/rpushx</remarks>
+        /// <remarks>
+        /// <seealso href="https://redis.io/commands/rpush"/>,
+        /// <seealso href="https://redis.io/commands/rpushx"/>
+        /// </remarks>
         long ListRightPush(RedisKey key, RedisValue[] values, When when = When.Always, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
         /// Insert all the specified values at the tail of the list stored at key.
         /// If key does not exist, it is created as empty list before performing the push operation.
         /// Elements are inserted one after the other to the tail of the list, from the leftmost element to the rightmost element.
-        /// So for instance the command RPUSH mylist a b c will result into a list containing a as first element, b as second element and c as third element.
+        /// So for instance the command <c>RPUSH mylist a b c</c> will result into a list containing a as first element, b as second element and c as third element.
         /// </summary>
         /// <param name="key">The key of the list.</param>
         /// <param name="values">The values to add to the tail of the list.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The length of the list after the push operation.</returns>
-        /// <remarks>https://redis.io/commands/rpush</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/rpush"/></remarks>
         long ListRightPush(RedisKey key, RedisValue[] values, CommandFlags flags);
 
         /// <summary>
@@ -1090,20 +1128,20 @@ namespace StackExchange.Redis
         /// <param name="index">The index to set the value at.</param>
         /// <param name="value">The values to add to the list.</param>
         /// <param name="flags">The flags to use for this operation.</param>
-        /// <remarks>https://redis.io/commands/lset</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/lset"/></remarks>
         void ListSetByIndex(RedisKey key, long index, RedisValue value, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
         /// Trim an existing list so that it will contain only the specified range of elements specified.
         /// Both start and stop are zero-based indexes, where 0 is the first element of the list (the head), 1 the next element and so on.
-        /// For example: LTRIM foobar 0 2 will modify the list stored at foobar so that only the first three elements of the list will remain.
+        /// For example: <c>LTRIM foobar 0 2</c> will modify the list stored at foobar so that only the first three elements of the list will remain.
         /// start and end can also be negative numbers indicating offsets from the end of the list, where -1 is the last element of the list, -2 the penultimate element and so on.
         /// </summary>
         /// <param name="key">The key of the list.</param>
         /// <param name="start">The start index of the list to trim to.</param>
         /// <param name="stop">The end index of the list to trim to.</param>
         /// <param name="flags">The flags to use for this operation.</param>
-        /// <remarks>https://redis.io/commands/ltrim</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/ltrim"/></remarks>
         void ListTrim(RedisKey key, long start, long stop, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -1153,7 +1191,7 @@ namespace StackExchange.Redis
         /// The number of clients that received the message *on the destination server*,
         /// note that this doesn't mean much in a cluster as clients can get the message through other nodes.
         /// </returns>
-        /// <remarks>https://redis.io/commands/publish</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/publish"/></remarks>
         long Publish(RedisChannel channel, RedisValue message, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -1162,8 +1200,8 @@ namespace StackExchange.Redis
         /// </summary>
         /// <param name="command">The command to run.</param>
         /// <param name="args">The arguments to pass for the command.</param>
-        /// <remarks>This API should be considered an advanced feature; inappropriate use can be harmful.</remarks>
         /// <returns>A dynamic representation of the command's result.</returns>
+        /// <remarks>This API should be considered an advanced feature; inappropriate use can be harmful.</remarks>
         RedisResult Execute(string command, params object[] args);
 
         /// <summary>
@@ -1173,8 +1211,8 @@ namespace StackExchange.Redis
         /// <param name="command">The command to run.</param>
         /// <param name="args">The arguments to pass for the command.</param>
         /// <param name="flags">The flags to use for this operation.</param>
-        /// <remarks>This API should be considered an advanced feature; inappropriate use can be harmful.</remarks>
         /// <returns>A dynamic representation of the command's result.</returns>
+        /// <remarks>This API should be considered an advanced feature; inappropriate use can be harmful.</remarks>
         RedisResult Execute(string command, ICollection<object> args, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -1185,8 +1223,10 @@ namespace StackExchange.Redis
         /// <param name="values">The values to execute against.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>A dynamic representation of the script's result.</returns>
-        /// <remarks>https://redis.io/commands/eval</remarks>
-        /// <remarks>https://redis.io/commands/evalsha</remarks>
+        /// <remarks>
+        /// <seealso href="https://redis.io/commands/eval"/>,
+        /// <seealso href="https://redis.io/commands/evalsha"/>
+        /// </remarks>
         RedisResult ScriptEvaluate(string script, RedisKey[]? keys = null, RedisValue[]? values = null, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -1197,7 +1237,7 @@ namespace StackExchange.Redis
         /// <param name="values">The values to execute against.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>A dynamic representation of the script's result.</returns>
-        /// <remarks>https://redis.io/commands/evalsha</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/evalsha"/></remarks>
         RedisResult ScriptEvaluate(byte[] hash, RedisKey[]? keys = null, RedisValue[]? values = null, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -1208,7 +1248,7 @@ namespace StackExchange.Redis
         /// <param name="parameters">The parameters to pass to the script.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>A dynamic representation of the script's result.</returns>
-        /// <remarks>https://redis.io/commands/eval</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/eval"/></remarks>
         RedisResult ScriptEvaluate(LuaScript script, object? parameters = null, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -1220,7 +1260,7 @@ namespace StackExchange.Redis
         /// <param name="parameters">The parameters to pass to the script.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>A dynamic representation of the script's result.</returns>
-        /// <remarks>https://redis.io/commands/eval</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/eval"/></remarks>
         RedisResult ScriptEvaluate(LoadedLuaScript script, object? parameters = null, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -1232,7 +1272,7 @@ namespace StackExchange.Redis
         /// <param name="value">The value to add to the set.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns><see langword="true"/> if the specified member was not already present in the set, else <see langword="false"/>.</returns>
-        /// <remarks>https://redis.io/commands/sadd</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/sadd"/></remarks>
         bool SetAdd(RedisKey key, RedisValue value, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -1244,7 +1284,7 @@ namespace StackExchange.Redis
         /// <param name="values">The values to add to the set.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The number of elements that were added to the set, not including all the elements already present into the set.</returns>
-        /// <remarks>https://redis.io/commands/sadd</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/sadd"/></remarks>
         long SetAdd(RedisKey key, RedisValue[] values, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -1255,9 +1295,11 @@ namespace StackExchange.Redis
         /// <param name="second">The key of the second set.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>List with members of the resulting set.</returns>
-        /// <remarks>https://redis.io/commands/sunion</remarks>
-        /// <remarks>https://redis.io/commands/sinter</remarks>
-        /// <remarks>https://redis.io/commands/sdiff</remarks>
+        /// <remarks>
+        /// <seealso href="https://redis.io/commands/sunion"/>,
+        /// <seealso href="https://redis.io/commands/sinter"/>,
+        /// <seealso href="https://redis.io/commands/sdiff"/>
+        /// </remarks>
         RedisValue[] SetCombine(SetOperation operation, RedisKey first, RedisKey second, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -1267,9 +1309,11 @@ namespace StackExchange.Redis
         /// <param name="keys">The keys of the sets to operate on.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>List with members of the resulting set.</returns>
-        /// <remarks>https://redis.io/commands/sunion</remarks>
-        /// <remarks>https://redis.io/commands/sinter</remarks>
-        /// <remarks>https://redis.io/commands/sdiff</remarks>
+        /// <remarks>
+        /// <seealso href="https://redis.io/commands/sunion"/>,
+        /// <seealso href="https://redis.io/commands/sinter"/>,
+        /// <seealso href="https://redis.io/commands/sdiff"/>
+        /// </remarks>
         RedisValue[] SetCombine(SetOperation operation, RedisKey[] keys, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -1282,9 +1326,11 @@ namespace StackExchange.Redis
         /// <param name="second">The key of the second set.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The number of elements in the resulting set.</returns>
-        /// <remarks>https://redis.io/commands/sunionstore</remarks>
-        /// <remarks>https://redis.io/commands/sinterstore</remarks>
-        /// <remarks>https://redis.io/commands/sdiffstore</remarks>
+        /// <remarks>
+        /// <seealso href="https://redis.io/commands/sunionstore"/>,
+        /// <seealso href="https://redis.io/commands/sinterstore"/>,
+        /// <seealso href="https://redis.io/commands/sdiffstore"/>
+        /// </remarks>
         long SetCombineAndStore(SetOperation operation, RedisKey destination, RedisKey first, RedisKey second, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -1296,9 +1342,11 @@ namespace StackExchange.Redis
         /// <param name="keys">The keys of the sets to operate on.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The number of elements in the resulting set.</returns>
-        /// <remarks>https://redis.io/commands/sunionstore</remarks>
-        /// <remarks>https://redis.io/commands/sinterstore</remarks>
-        /// <remarks>https://redis.io/commands/sdiffstore</remarks>
+        /// <remarks>
+        /// <seealso href="https://redis.io/commands/sunionstore"/>,
+        /// <seealso href="https://redis.io/commands/sinterstore"/>,
+        /// <seealso href="https://redis.io/commands/sdiffstore"/>
+        /// </remarks>
         long SetCombineAndStore(SetOperation operation, RedisKey destination, RedisKey[] keys, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -1311,7 +1359,7 @@ namespace StackExchange.Redis
         /// <see langword="true"/> if the element is a member of the set.
         /// <see langword="false"/> if the element is not a member of the set, or if key does not exist.
         /// </returns>
-        /// <remarks>https://redis.io/commands/sismember</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/sismember"/></remarks>
         bool SetContains(RedisKey key, RedisValue value, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -1321,11 +1369,10 @@ namespace StackExchange.Redis
         /// <param name="values">The members to check for.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>
-        /// An array of booleans corresponding to <paramref name="values"/>, for each:
         /// <see langword="true"/> if the element is a member of the set.
         /// <see langword="false"/> if the element is not a member of the set, or if key does not exist.
         /// </returns>
-        /// <remarks>https://redis.io/commands/smismember</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/smismember"/></remarks>
         bool[] SetContains(RedisKey key, RedisValue[] values, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -1341,7 +1388,7 @@ namespace StackExchange.Redis
         /// <param name="limit">The number of elements to check (defaults to 0 and means unlimited).</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The cardinality (number of elements) of the set, or 0 if key does not exist.</returns>
-        /// <remarks>https://redis.io/commands/scard</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/scard"/></remarks>
         long SetIntersectionLength(RedisKey[] keys, long limit = 0, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -1350,7 +1397,7 @@ namespace StackExchange.Redis
         /// <param name="key">The key of the set.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The cardinality (number of elements) of the set, or 0 if key does not exist.</returns>
-        /// <remarks>https://redis.io/commands/scard</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/scard"/></remarks>
         long SetLength(RedisKey key, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -1359,7 +1406,7 @@ namespace StackExchange.Redis
         /// <param name="key">The key of the set.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>All elements of the set.</returns>
-        /// <remarks>https://redis.io/commands/smembers</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/smembers"/></remarks>
         RedisValue[] SetMembers(RedisKey key, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -1375,7 +1422,7 @@ namespace StackExchange.Redis
         /// <see langword="true"/> if the element is moved.
         /// <see langword="false"/> if the element is not a member of source and no operation was performed.
         /// </returns>
-        /// <remarks>https://redis.io/commands/smove</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/smove"/></remarks>
         bool SetMove(RedisKey source, RedisKey destination, RedisValue value, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -1384,7 +1431,7 @@ namespace StackExchange.Redis
         /// <param name="key">The key of the set.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The removed element, or nil when key does not exist.</returns>
-        /// <remarks>https://redis.io/commands/spop</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/spop"/></remarks>
         RedisValue SetPop(RedisKey key, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -1394,7 +1441,7 @@ namespace StackExchange.Redis
         /// <param name="count">The number of elements to return.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>An array of elements, or an empty array when key does not exist.</returns>
-        /// <remarks>https://redis.io/commands/spop</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/spop"/></remarks>
         RedisValue[] SetPop(RedisKey key, long count, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -1403,7 +1450,7 @@ namespace StackExchange.Redis
         /// <param name="key">The key of the set.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The randomly selected element, or <see cref="RedisValue.Null"/> when <paramref name="key"/> does not exist.</returns>
-        /// <remarks>https://redis.io/commands/srandmember</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/srandmember"/></remarks>
         RedisValue SetRandomMember(RedisKey key, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -1415,7 +1462,7 @@ namespace StackExchange.Redis
         /// <param name="count">The count of members to get.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>An array of elements, or an empty array when <paramref name="key"/> does not exist.</returns>
-        /// <remarks>https://redis.io/commands/srandmember</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/srandmember"/></remarks>
         RedisValue[] SetRandomMembers(RedisKey key, long count, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -1426,7 +1473,7 @@ namespace StackExchange.Redis
         /// <param name="value">The value to remove.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns><see langword="true"/> if the specified member was already present in the set, <see langword="false"/> otherwise.</returns>
-        /// <remarks>https://redis.io/commands/srem</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/srem"/></remarks>
         bool SetRemove(RedisKey key, RedisValue value, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -1437,7 +1484,7 @@ namespace StackExchange.Redis
         /// <param name="values">The values to remove.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The number of members that were removed from the set, not including non existing members.</returns>
-        /// <remarks>https://redis.io/commands/srem</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/srem"/></remarks>
         long SetRemove(RedisKey key, RedisValue[] values, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -1448,7 +1495,7 @@ namespace StackExchange.Redis
         /// <param name="pageSize">The page size to iterate by.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>Yields all matching elements of the set.</returns>
-        /// <remarks>https://redis.io/commands/sscan</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/sscan"/></remarks>
         IEnumerable<RedisValue> SetScan(RedisKey key, RedisValue pattern, int pageSize, CommandFlags flags);
 
         /// <summary>
@@ -1462,7 +1509,7 @@ namespace StackExchange.Redis
         /// <param name="pageOffset">The page offset to start at.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>Yields all matching elements of the set.</returns>
-        /// <remarks>https://redis.io/commands/sscan</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/sscan"/></remarks>
         IEnumerable<RedisValue> SetScan(RedisKey key, RedisValue pattern = default, int pageSize = RedisBase.CursorUtils.DefaultLibraryPageSize, long cursor = RedisBase.CursorUtils.Origin, int pageOffset = 0, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -1482,7 +1529,7 @@ namespace StackExchange.Redis
         /// <param name="get">The key pattern to sort by, if any e.g. ExternalKey_* would return the value of ExternalKey_{listvalue} for each entry.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The sorted elements, or the external values if <c>get</c> is specified.</returns>
-        /// <remarks>https://redis.io/commands/sort</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/sort"/></remarks>
         RedisValue[] Sort(RedisKey key, long skip = 0, long take = -1, Order order = Order.Ascending, SortType sortType = SortType.Numeric, RedisValue by = default, RedisValue[]? get = null, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -1490,7 +1537,7 @@ namespace StackExchange.Redis
         /// By default, the elements themselves are compared, but the values can also be used to perform external key-lookups using the <c>by</c> parameter.
         /// By default, the elements themselves are returned, but external key-lookups (one or many) can be performed instead by specifying
         /// the <c>get</c> parameter (note that <c>#</c> specifies the element itself, when used in <c>get</c>).
-        /// Referring to the <a href="https://redis.io/commands/sort">redis SORT documentation </a> for examples is recommended.
+        /// Referring to the <a href="https://redis.io/commands/sort">redis SORT documentation</a> for examples is recommended.
         /// When used in hashes, <c>by</c> and <c>get</c> can be used to specify fields using <c>-&gt;</c> notation (again, refer to redis documentation).
         /// </summary>
         /// <param name="destination">The destination key to store results in.</param>
@@ -1503,7 +1550,7 @@ namespace StackExchange.Redis
         /// <param name="get">The key pattern to sort by, if any e.g. ExternalKey_* would return the value of ExternalKey_{listvalue} for each entry.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The number of elements stored in the new list.</returns>
-        /// <remarks>https://redis.io/commands/sort</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/sort"/></remarks>
         long SortAndStore(RedisKey destination, RedisKey key, long skip = 0, long take = -1, Order order = Order.Ascending, SortType sortType = SortType.Numeric, RedisValue by = default, RedisValue[]? get = null, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -1515,7 +1562,7 @@ namespace StackExchange.Redis
         /// <param name="score">The score for the member to add to the sorted set.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns><see langword="true"/> if the value was added. <see langword="false"/> if it already existed (the score is still updated).</returns>
-        /// <remarks>https://redis.io/commands/zadd</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/zadd"/></remarks>
         bool SortedSetAdd(RedisKey key, RedisValue member, double score, CommandFlags flags);
 
         /// <summary>
@@ -1528,7 +1575,7 @@ namespace StackExchange.Redis
         /// <param name="when">What conditions to add the element under (defaults to always).</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns><see langword="true"/> if the value was added. <see langword="false"/> if it already existed (the score is still updated).</returns>
-        /// <remarks>https://redis.io/commands/zadd</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/zadd"/></remarks>
         bool SortedSetAdd(RedisKey key, RedisValue member, double score, When when = When.Always, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -1539,7 +1586,7 @@ namespace StackExchange.Redis
         /// <param name="values">The members and values to add to the sorted set.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The number of elements added to the sorted sets, not including elements already existing for which the score was updated.</returns>
-        /// <remarks>https://redis.io/commands/zadd</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/zadd"/></remarks>
         long SortedSetAdd(RedisKey key, SortedSetEntry[] values, CommandFlags flags);
 
         /// <summary>
@@ -1551,7 +1598,7 @@ namespace StackExchange.Redis
         /// <param name="when">What conditions to add the element under (defaults to always).</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The number of elements added to the sorted sets, not including elements already existing for which the score was updated.</returns>
-        /// <remarks>https://redis.io/commands/zadd</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/zadd"/></remarks>
         long SortedSetAdd(RedisKey key, SortedSetEntry[] values, When when = When.Always, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -1564,10 +1611,12 @@ namespace StackExchange.Redis
         /// <param name="weights">The optional weights per set that correspond to <paramref name="keys"/>.</param>
         /// <param name="aggregate">The aggregation method (defaults to <see cref="Aggregate.Sum"/>).</param>
         /// <param name="flags">The flags to use for this operation.</param>
-        /// <remarks>https://redis.io/commands/zunion</remarks>
-        /// <remarks>https://redis.io/commands/zinter</remarks>
-        /// <remarks>https://redis.io/commands/zdiff</remarks>
         /// <returns>The resulting sorted set.</returns>
+        /// <remarks>
+        /// <seealso href="https://redis.io/commands/zunion"/>,
+        /// <seealso href="https://redis.io/commands/zinter"/>,
+        /// <seealso href="https://redis.io/commands/zdiff"/>
+        /// </remarks>
         RedisValue[] SortedSetCombine(SetOperation operation, RedisKey[] keys, double[]? weights = null, Aggregate aggregate = Aggregate.Sum, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -1580,10 +1629,12 @@ namespace StackExchange.Redis
         /// <param name="weights">The optional weights per set that correspond to <paramref name="keys"/>.</param>
         /// <param name="aggregate">The aggregation method (defaults to <see cref="Aggregate.Sum"/>).</param>
         /// <param name="flags">The flags to use for this operation.</param>
-        /// <remarks>https://redis.io/commands/zunion</remarks>
-        /// <remarks>https://redis.io/commands/zinter</remarks>
-        /// <remarks>https://redis.io/commands/zdiff</remarks>
         /// <returns>The resulting sorted set with scores.</returns>
+        /// <remarks>
+        /// <seealso href="https://redis.io/commands/zunion"/>,
+        /// <seealso href="https://redis.io/commands/zinter"/>,
+        /// <seealso href="https://redis.io/commands/zdiff"/>
+        /// </remarks>
         SortedSetEntry[] SortedSetCombineWithScores(SetOperation operation, RedisKey[] keys, double[]? weights = null, Aggregate aggregate = Aggregate.Sum, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -1597,10 +1648,12 @@ namespace StackExchange.Redis
         /// <param name="second">The key of the second sorted set.</param>
         /// <param name="aggregate">The aggregation method (defaults to sum).</param>
         /// <param name="flags">The flags to use for this operation.</param>
-        /// <remarks>https://redis.io/commands/zunionstore</remarks>
-        /// <remarks>https://redis.io/commands/zinterstore</remarks>
-        /// <remarks>https://redis.io/commands/zdiffstore</remarks>
         /// <returns>The number of elements in the resulting sorted set at destination.</returns>
+        /// <remarks>
+        /// <seealso href="https://redis.io/commands/zunionstore"/>,
+        /// <seealso href="https://redis.io/commands/zinterstore"/>,
+        /// <seealso href="https://redis.io/commands/zdiffstore"/>
+        /// </remarks>
         long SortedSetCombineAndStore(SetOperation operation, RedisKey destination, RedisKey first, RedisKey second, Aggregate aggregate = Aggregate.Sum, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -1614,10 +1667,12 @@ namespace StackExchange.Redis
         /// <param name="weights">The optional weights per set that correspond to <paramref name="keys"/>.</param>
         /// <param name="aggregate">The aggregation method (defaults to sum).</param>
         /// <param name="flags">The flags to use for this operation.</param>
-        /// <remarks>https://redis.io/commands/zunionstore</remarks>
-        /// <remarks>https://redis.io/commands/zinterstore</remarks>
-        /// <remarks>https://redis.io/commands/zdiffstore</remarks>
         /// <returns>The number of elements in the resulting sorted set at destination.</returns>
+        /// <remarks>
+        /// <seealso href="https://redis.io/commands/zunionstore"/>,
+        /// <seealso href="https://redis.io/commands/zinterstore"/>,
+        /// <seealso href="https://redis.io/commands/zdiffstore"/>
+        /// </remarks>
         long SortedSetCombineAndStore(SetOperation operation, RedisKey destination, RedisKey[] keys, double[]? weights = null, Aggregate aggregate = Aggregate.Sum, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -1629,7 +1684,7 @@ namespace StackExchange.Redis
         /// <param name="value">The amount to decrement by.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The new score of member.</returns>
-        /// <remarks>https://redis.io/commands/zincrby</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/zincrby"/></remarks>
         double SortedSetDecrement(RedisKey key, RedisValue member, double value, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -1640,7 +1695,7 @@ namespace StackExchange.Redis
         /// <param name="value">The amount to increment by.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The new score of member.</returns>
-        /// <remarks>https://redis.io/commands/zincrby</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/zincrby"/></remarks>
         double SortedSetIncrement(RedisKey key, RedisValue member, double value, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -1650,7 +1705,7 @@ namespace StackExchange.Redis
         /// <param name="limit">If the intersection cardinality reaches <paramref name="limit"/> partway through the computation, the algorithm will exit and yield <paramref name="limit"/> as the cardinality (defaults to 0 meaning unlimited).</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The number of elements in the resulting intersection.</returns>
-        /// <remarks>https://redis.io/commands/zintercard</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/zintercard"/></remarks>
         long SortedSetIntersectionLength(RedisKey[] keys, long limit = 0, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -1662,7 +1717,7 @@ namespace StackExchange.Redis
         /// <param name="exclude">Whether to exclude <paramref name="min"/> and <paramref name="max"/> from the range check (defaults to both inclusive).</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The cardinality (number of elements) of the sorted set, or 0 if key does not exist.</returns>
-        /// <remarks>https://redis.io/commands/zcard</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/zcard"/></remarks>
         long SortedSetLength(RedisKey key, double min = double.NegativeInfinity, double max = double.PositiveInfinity, Exclude exclude = Exclude.None, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -1675,7 +1730,7 @@ namespace StackExchange.Redis
         /// <param name="exclude">Whether to exclude <paramref name="min"/> and <paramref name="max"/> from the range check (defaults to both inclusive).</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The number of elements in the specified score range.</returns>
-        /// <remarks>https://redis.io/commands/zlexcount</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/zlexcount"/></remarks>
         long SortedSetLengthByValue(RedisKey key, RedisValue min, RedisValue max, Exclude exclude = Exclude.None, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -1684,7 +1739,7 @@ namespace StackExchange.Redis
         /// <param name="key">The key of the sorted set.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The randomly selected element, or <see cref="RedisValue.Null"/> when <paramref name="key"/> does not exist.</returns>
-        /// <remarks>https://redis.io/commands/zrandmember</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/zrandmember"/></remarks>
         RedisValue SortedSetRandomMember(RedisKey key, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -1703,7 +1758,7 @@ namespace StackExchange.Redis
         /// </param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The randomly selected elements, or an empty array when <paramref name="key"/> does not exist.</returns>
-        /// <remarks>https://redis.io/commands/zrandmember</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/zrandmember"/></remarks>
         RedisValue[] SortedSetRandomMembers(RedisKey key, long count, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -1722,7 +1777,7 @@ namespace StackExchange.Redis
         /// </param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The randomly selected elements with scores, or an empty array when <paramref name="key"/> does not exist.</returns>
-        /// <remarks>https://redis.io/commands/zrandmember</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/zrandmember"/></remarks>
         SortedSetEntry[] SortedSetRandomMembersWithScores(RedisKey key, long count, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -1738,8 +1793,10 @@ namespace StackExchange.Redis
         /// <param name="order">The order to sort by (defaults to ascending).</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>List of elements in the specified range.</returns>
-        /// <remarks>https://redis.io/commands/zrange</remarks>
-        /// <remarks>https://redis.io/commands/zrevrange</remarks>
+        /// <remarks>
+        /// <seealso href="https://redis.io/commands/zrange"/>,
+        /// <seealso href="https://redis.io/commands/zrevrange"/>
+        /// </remarks>
         RedisValue[] SortedSetRangeByRank(RedisKey key, long start = 0, long stop = -1, Order order = Order.Ascending, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -1760,8 +1817,8 @@ namespace StackExchange.Redis
         /// <param name="skip">The number of elements into the sorted set to skip. Note: this iterates after sorting so incurs O(n) cost for large values.</param>
         /// <param name="take">The maximum number of elements to pull into the new (<paramref name="destinationKey"/>) set.</param>
         /// <param name="flags">The flags to use for this operation.</param>
-        /// <remarks>https://redis.io/commands/zrangestore</remarks>
         /// <returns>The cardinality of (number of elements in) the newly created sorted set.</returns>
+        /// <remarks><seealso href="https://redis.io/commands/zrangestore"/></remarks>
         long SortedSetRangeAndStore(
             RedisKey sourceKey,
             RedisKey destinationKey,
@@ -1787,8 +1844,10 @@ namespace StackExchange.Redis
         /// <param name="order">The order to sort by (defaults to ascending).</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>List of elements in the specified range.</returns>
-        /// <remarks>https://redis.io/commands/zrange</remarks>
-        /// <remarks>https://redis.io/commands/zrevrange</remarks>
+        /// <remarks>
+        /// <seealso href="https://redis.io/commands/zrange"/>,
+        /// <seealso href="https://redis.io/commands/zrevrange"/>
+        /// </remarks>
         SortedSetEntry[] SortedSetRangeByRankWithScores(RedisKey key, long start = 0, long stop = -1, Order order = Order.Ascending, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -1807,8 +1866,10 @@ namespace StackExchange.Redis
         /// <param name="take">How many items to take.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>List of elements in the specified score range.</returns>
-        /// <remarks>https://redis.io/commands/zrangebyscore</remarks>
-        /// <remarks>https://redis.io/commands/zrevrangebyscore</remarks>
+        /// <remarks>
+        /// <seealso href="https://redis.io/commands/zrangebyscore"/>,
+        /// <seealso href="https://redis.io/commands/zrevrangebyscore"/>
+        /// </remarks>
         RedisValue[] SortedSetRangeByScore(RedisKey key,
             double start = double.NegativeInfinity,
             double stop = double.PositiveInfinity,
@@ -1834,8 +1895,10 @@ namespace StackExchange.Redis
         /// <param name="take">How many items to take.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>List of elements in the specified score range.</returns>
-        /// <remarks>https://redis.io/commands/zrangebyscore</remarks>
-        /// <remarks>https://redis.io/commands/zrevrangebyscore</remarks>
+        /// <remarks>
+        /// <seealso href="https://redis.io/commands/zrangebyscore"/>,
+        /// <seealso href="https://redis.io/commands/zrevrangebyscore"/>
+        /// </remarks>
         SortedSetEntry[] SortedSetRangeByScoreWithScores(RedisKey key,
             double start = double.NegativeInfinity,
             double stop = double.PositiveInfinity,
@@ -1856,8 +1919,8 @@ namespace StackExchange.Redis
         /// <param name="skip">How many items to skip.</param>
         /// <param name="take">How many items to take.</param>
         /// <param name="flags">The flags to use for this operation.</param>
-        /// <remarks>https://redis.io/commands/zrangebylex</remarks>
         /// <returns>List of elements in the specified score range.</returns>
+        /// <remarks><seealso href="https://redis.io/commands/zrangebylex"/></remarks>
         RedisValue[] SortedSetRangeByValue(RedisKey key,
             RedisValue min,
             RedisValue max,
@@ -1878,9 +1941,11 @@ namespace StackExchange.Redis
         /// <param name="skip">How many items to skip.</param>
         /// <param name="take">How many items to take.</param>
         /// <param name="flags">The flags to use for this operation.</param>
-        /// <remarks>https://redis.io/commands/zrangebylex</remarks>
-        /// <remarks>https://redis.io/commands/zrevrangebylex</remarks>
         /// <returns>List of elements in the specified score range.</returns>
+        /// <remarks>
+        /// <seealso href="https://redis.io/commands/zrangebylex"/>,
+        /// <seealso href="https://redis.io/commands/zrevrangebylex"/>
+        /// </remarks>
         RedisValue[] SortedSetRangeByValue(RedisKey key,
             RedisValue min = default,
             RedisValue max = default,
@@ -1899,8 +1964,10 @@ namespace StackExchange.Redis
         /// <param name="order">The order to sort by (defaults to ascending).</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>If member exists in the sorted set, the rank of member. If member does not exist in the sorted set or key does not exist, <see langword="null"/>.</returns>
-        /// <remarks>https://redis.io/commands/zrank</remarks>
-        /// <remarks>https://redis.io/commands/zrevrank</remarks>
+        /// <remarks>
+        /// <seealso href="https://redis.io/commands/zrank"/>,
+        /// <seealso href="https://redis.io/commands/zrevrank"/>
+        /// </remarks>
         long? SortedSetRank(RedisKey key, RedisValue member, Order order = Order.Ascending, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -1910,7 +1977,7 @@ namespace StackExchange.Redis
         /// <param name="member">The member to remove.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns><see langword="true"/> if the member existed in the sorted set and was removed. <see langword="false"/> otherwise.</returns>
-        /// <remarks>https://redis.io/commands/zrem</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/zrem"/></remarks>
         bool SortedSetRemove(RedisKey key, RedisValue member, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -1920,7 +1987,7 @@ namespace StackExchange.Redis
         /// <param name="members">The members to remove.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The number of members removed from the sorted set, not including non existing members.</returns>
-        /// <remarks>https://redis.io/commands/zrem</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/zrem"/></remarks>
         long SortedSetRemove(RedisKey key, RedisValue[] members, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -1934,7 +2001,7 @@ namespace StackExchange.Redis
         /// <param name="stop">The maximum rank to remove.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The number of elements removed.</returns>
-        /// <remarks>https://redis.io/commands/zremrangebyrank</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/zremrangebyrank"/></remarks>
         long SortedSetRemoveRangeByRank(RedisKey key, long start, long stop, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -1946,7 +2013,7 @@ namespace StackExchange.Redis
         /// <param name="exclude">Which of <paramref name="start"/> and <paramref name="stop"/> to exclude (defaults to both inclusive).</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The number of elements removed.</returns>
-        /// <remarks>https://redis.io/commands/zremrangebyscore</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/zremrangebyscore"/></remarks>
         long SortedSetRemoveRangeByScore(RedisKey key, double start, double stop, Exclude exclude = Exclude.None, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -1959,7 +2026,7 @@ namespace StackExchange.Redis
         /// <param name="exclude">Which of <paramref name="min"/> and <paramref name="max"/> to exclude (defaults to both inclusive).</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The number of elements removed.</returns>
-        /// <remarks>https://redis.io/commands/zremrangebylex</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/zremrangebylex"/></remarks>
         long SortedSetRemoveRangeByValue(RedisKey key, RedisValue min, RedisValue max, Exclude exclude = Exclude.None, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -1970,7 +2037,7 @@ namespace StackExchange.Redis
         /// <param name="pageSize">The page size to iterate by.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>Yields all matching elements of the sorted set.</returns>
-        /// <remarks>https://redis.io/commands/zscan</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/zscan"/></remarks>
         IEnumerable<SortedSetEntry> SortedSetScan(RedisKey key, RedisValue pattern, int pageSize, CommandFlags flags);
 
         /// <summary>
@@ -1984,7 +2051,7 @@ namespace StackExchange.Redis
         /// <param name="pageOffset">The page offset to start at.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>Yields all matching elements of the sorted set.</returns>
-        /// <remarks>https://redis.io/commands/zscan</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/zscan"/></remarks>
         IEnumerable<SortedSetEntry> SortedSetScan(RedisKey key,
             RedisValue pattern = default,
             int pageSize = RedisBase.CursorUtils.DefaultLibraryPageSize,
@@ -2000,7 +2067,7 @@ namespace StackExchange.Redis
         /// <param name="member">The member to get a score for.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The score of the member.</returns>
-        /// <remarks>https://redis.io/commands/zscore</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/zscore"/></remarks>
         double? SortedSetScore(RedisKey key, RedisValue member, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -2014,7 +2081,7 @@ namespace StackExchange.Redis
         /// The scores of the members in the same order as the <paramref name="members"/> array.
         /// If a member does not exist in the set, <see langword="null"/> is returned.
         /// </returns>
-        /// <remarks>https://redis.io/commands/zmscore</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/zmscore"/></remarks>
         double?[] SortedSetScores(RedisKey key, RedisValue[] members, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -2024,8 +2091,10 @@ namespace StackExchange.Redis
         /// <param name="order">The order to sort by (defaults to ascending).</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The removed element, or nil when key does not exist.</returns>
-        /// <remarks>https://redis.io/commands/zpopmin</remarks>
-        /// <remarks>https://redis.io/commands/zpopmax</remarks>
+        /// <remarks>
+        /// <seealso href="https://redis.io/commands/zpopmin"/>,
+        /// <seealso href="https://redis.io/commands/zpopmax"/>
+        /// </remarks>
         SortedSetEntry? SortedSetPop(RedisKey key, Order order = Order.Ascending, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -2036,8 +2105,10 @@ namespace StackExchange.Redis
         /// <param name="order">The order to sort by (defaults to ascending).</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>An array of elements, or an empty array when key does not exist.</returns>
-        /// <remarks>https://redis.io/commands/zpopmin</remarks>
-        /// <remarks>https://redis.io/commands/zpopmax</remarks>
+        /// <remarks>
+        /// <seealso href="https://redis.io/commands/zpopmin"/>,
+        /// <seealso href="https://redis.io/commands/zpopmax"/>
+        /// </remarks>
         SortedSetEntry[] SortedSetPop(RedisKey key, long count, Order order = Order.Ascending, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -2048,7 +2119,7 @@ namespace StackExchange.Redis
         /// <param name="messageId">The ID of the message to acknowledge.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The number of messages acknowledged.</returns>
-        /// <remarks>https://redis.io/topics/streams-intro</remarks>
+        /// <remarks><seealso href="https://redis.io/topics/streams-intro"/></remarks>
         long StreamAcknowledge(RedisKey key, RedisValue groupName, RedisValue messageId, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -2059,7 +2130,7 @@ namespace StackExchange.Redis
         /// <param name="messageIds">The IDs of the messages to acknowledge.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The number of messages acknowledged.</returns>
-        /// <remarks>https://redis.io/topics/streams-intro</remarks>
+        /// <remarks><seealso href="https://redis.io/topics/streams-intro"/></remarks>
         long StreamAcknowledge(RedisKey key, RedisValue groupName, RedisValue[] messageIds, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -2075,7 +2146,7 @@ namespace StackExchange.Redis
         /// <param name="useApproximateMaxLength">If true, the "~" argument is used to allow the stream to exceed max length by a small number. This improves performance when removing messages.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The ID of the newly created message.</returns>
-        /// <remarks>https://redis.io/commands/xadd</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/xadd"/></remarks>
         RedisValue StreamAdd(RedisKey key, RedisValue streamField, RedisValue streamValue, RedisValue? messageId = null, int? maxLength = null, bool useApproximateMaxLength = false, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -2090,7 +2161,7 @@ namespace StackExchange.Redis
         /// <param name="useApproximateMaxLength">If true, the "~" argument is used to allow the stream to exceed max length by a small number. This improves performance when removing messages.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The ID of the newly created message.</returns>
-        /// <remarks>https://redis.io/commands/xadd</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/xadd"/></remarks>
         RedisValue StreamAdd(RedisKey key, NameValueEntry[] streamPairs, RedisValue? messageId = null, int? maxLength = null, bool useApproximateMaxLength = false, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -2104,7 +2175,7 @@ namespace StackExchange.Redis
         /// <param name="messageIds">The IDs of the messages to claim for the given consumer.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The messages successfully claimed by the given consumer.</returns>
-        /// <remarks>https://redis.io/topics/streams-intro</remarks>
+        /// <remarks><seealso href="https://redis.io/topics/streams-intro"/></remarks>
         StreamEntry[] StreamClaim(RedisKey key, RedisValue consumerGroup, RedisValue claimingConsumer, long minIdleTimeInMs, RedisValue[] messageIds, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -2118,7 +2189,7 @@ namespace StackExchange.Redis
         /// <param name="messageIds">The IDs of the messages to claim for the given consumer.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The message IDs for the messages successfully claimed by the given consumer.</returns>
-        /// <remarks>https://redis.io/topics/streams-intro</remarks>
+        /// <remarks><seealso href="https://redis.io/topics/streams-intro"/></remarks>
         RedisValue[] StreamClaimIdsOnly(RedisKey key, RedisValue consumerGroup, RedisValue claimingConsumer, long minIdleTimeInMs, RedisValue[] messageIds, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -2129,6 +2200,7 @@ namespace StackExchange.Redis
         /// <param name="position">The position from which to read for the consumer group.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns><see langword="true"/> if successful, <see langword="false"/> otherwise.</returns>
+        /// <remarks><seealso href="https://redis.io/topics/streams-intro"/></remarks>
         bool StreamConsumerGroupSetPosition(RedisKey key, RedisValue groupName, RedisValue position, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -2139,7 +2211,7 @@ namespace StackExchange.Redis
         /// <param name="groupName">The consumer group name.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>An instance of <see cref="StreamConsumerInfo"/> for each of the consumer group's consumers.</returns>
-        /// <remarks>https://redis.io/topics/streams-intro</remarks>
+        /// <remarks><seealso href="https://redis.io/topics/streams-intro"/></remarks>
         StreamConsumerInfo[] StreamConsumerInfo(RedisKey key, RedisValue groupName, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -2150,7 +2222,7 @@ namespace StackExchange.Redis
         /// <param name="position">The position to begin reading the stream. Defaults to <see cref="StreamPosition.NewMessages"/>.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns><see langword="true"/> if the group was created, <see langword="false"/> otherwise.</returns>
-        /// <remarks>https://redis.io/topics/streams-intro</remarks>
+        /// <remarks><seealso href="https://redis.io/topics/streams-intro"/></remarks>
         bool StreamCreateConsumerGroup(RedisKey key, RedisValue groupName, RedisValue? position, CommandFlags flags);
 
         /// <summary>
@@ -2162,7 +2234,7 @@ namespace StackExchange.Redis
         /// <param name="createStream">Create the stream if it does not already exist.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns><see langword="true"/> if the group was created, <see langword="false"/> otherwise.</returns>
-        /// <remarks>https://redis.io/topics/streams-intro</remarks>
+        /// <remarks><seealso href="https://redis.io/topics/streams-intro"/></remarks>
         bool StreamCreateConsumerGroup(RedisKey key, RedisValue groupName, RedisValue? position = null, bool createStream = true, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -2172,7 +2244,7 @@ namespace StackExchange.Redis
         /// <param name="messageIds">The IDs of the messages to delete.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>Returns the number of messages successfully deleted from the stream.</returns>
-        /// <remarks>https://redis.io/topics/streams-intro</remarks>
+        /// <remarks><seealso href="https://redis.io/topics/streams-intro"/></remarks>
         long StreamDelete(RedisKey key, RedisValue[] messageIds, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -2183,6 +2255,7 @@ namespace StackExchange.Redis
         /// <param name="consumerName">The name of the consumer.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The number of messages that were pending for the deleted consumer.</returns>
+        /// <remarks><seealso href="https://redis.io/topics/streams-intro"/></remarks>
         long StreamDeleteConsumer(RedisKey key, RedisValue groupName, RedisValue consumerName, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -2192,6 +2265,7 @@ namespace StackExchange.Redis
         /// <param name="groupName">The name of the consumer group.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns><see langword="true"/> if deleted, <see langword="false"/> otherwise.</returns>
+        /// <remarks><seealso href="https://redis.io/topics/streams-intro"/></remarks>
         bool StreamDeleteConsumerGroup(RedisKey key, RedisValue groupName, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -2200,7 +2274,7 @@ namespace StackExchange.Redis
         /// <param name="key">The key of the stream.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>An instance of <see cref="StreamGroupInfo"/> for each of the stream's groups.</returns>
-        /// <remarks>https://redis.io/topics/streams-intro</remarks>
+        /// <remarks><seealso href="https://redis.io/topics/streams-intro"/></remarks>
         StreamGroupInfo[] StreamGroupInfo(RedisKey key, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -2209,7 +2283,7 @@ namespace StackExchange.Redis
         /// <param name="key">The key of the stream.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>A <see cref="StreamInfo"/> instance with information about the stream.</returns>
-        /// <remarks>https://redis.io/topics/streams-intro</remarks>
+        /// <remarks><seealso href="https://redis.io/topics/streams-intro"/></remarks>
         StreamInfo StreamInfo(RedisKey key, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -2218,7 +2292,7 @@ namespace StackExchange.Redis
         /// <param name="key">The key of the stream.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The number of entries inside the given stream.</returns>
-        /// <remarks>https://redis.io/commands/xlen</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/xlen"/></remarks>
         long StreamLength(RedisKey key, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -2234,7 +2308,7 @@ namespace StackExchange.Redis
         /// The highest and lowest ID of the pending messages, and the consumers with their pending message count.
         /// </returns>
         /// <remarks>The equivalent of calling XPENDING key group.</remarks>
-        /// <remarks>https://redis.io/commands/xpending</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/xpending"/></remarks>
         StreamPendingInfo StreamPending(RedisKey key, RedisValue groupName, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -2249,7 +2323,7 @@ namespace StackExchange.Redis
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>An instance of <see cref="StreamPendingMessageInfo"/> for each pending message.</returns>
         /// <remarks>Equivalent of calling XPENDING key group start-id end-id count consumer-name.</remarks>
-        /// <remarks>https://redis.io/commands/xpending</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/xpending"/></remarks>
         StreamPendingMessageInfo[] StreamPendingMessages(RedisKey key, RedisValue groupName, int count, RedisValue consumerName, RedisValue? minId = null, RedisValue? maxId = null, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -2262,7 +2336,7 @@ namespace StackExchange.Redis
         /// <param name="messageOrder">The order of the messages. <see cref="Order.Ascending"/> will execute XRANGE and <see cref="Order.Descending"/> will execute XREVRANGE.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>Returns an instance of <see cref="StreamEntry"/> for each message returned.</returns>
-        /// <remarks>https://redis.io/commands/xrange</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/xrange"/></remarks>
         StreamEntry[] StreamRange(RedisKey key, RedisValue? minId = null, RedisValue? maxId = null, int? count = null, Order messageOrder = Order.Ascending, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -2273,8 +2347,10 @@ namespace StackExchange.Redis
         /// <param name="count">The maximum number of messages to return.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>Returns an instance of <see cref="StreamEntry"/> for each message returned.</returns>
-        /// <remarks>Equivalent of calling XREAD COUNT num STREAMS key id.</remarks>
-        /// <remarks>https://redis.io/commands/xread</remarks>
+        /// <remarks>
+        /// <para>Equivalent of calling <c>XREAD COUNT num STREAMS key id</c>.</para>
+        /// <para><seealso href="https://redis.io/commands/xread"/></para>
+        /// </remarks>
         StreamEntry[] StreamRead(RedisKey key, RedisValue position, int? count = null, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -2284,8 +2360,10 @@ namespace StackExchange.Redis
         /// <param name="countPerStream">The maximum number of messages to return from each stream.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>A value of <see cref="RedisStream"/> for each stream.</returns>
-        /// <remarks>Equivalent of calling XREAD COUNT num STREAMS key1 key2 id1 id2.</remarks>
-        /// <remarks>https://redis.io/commands/xread</remarks>
+        /// <remarks>
+        /// <para>Equivalent of calling <c>XREAD COUNT num STREAMS key1 key2 id1 id2</c>.</para>
+        /// <para><seealso href="https://redis.io/commands/xread"/></para>
+        /// </remarks>
         RedisStream[] StreamRead(StreamPosition[] streamPositions, int? countPerStream = null, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -2298,7 +2376,7 @@ namespace StackExchange.Redis
         /// <param name="count">The maximum number of messages to return.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>Returns a value of <see cref="StreamEntry"/> for each message returned.</returns>
-        /// <remarks>https://redis.io/commands/xreadgroup</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/xreadgroup"/></remarks>
         StreamEntry[] StreamReadGroup(RedisKey key, RedisValue groupName, RedisValue consumerName, RedisValue? position, int? count, CommandFlags flags);
 
         /// <summary>
@@ -2312,7 +2390,7 @@ namespace StackExchange.Redis
         /// <param name="noAck">When true, the message will not be added to the pending message list.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>Returns a value of <see cref="StreamEntry"/> for each message returned.</returns>
-        /// <remarks>https://redis.io/commands/xreadgroup</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/xreadgroup"/></remarks>
         StreamEntry[] StreamReadGroup(RedisKey key, RedisValue groupName, RedisValue consumerName, RedisValue? position = null, int? count = null, bool noAck = false, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -2325,8 +2403,10 @@ namespace StackExchange.Redis
         /// <param name="countPerStream">The maximum number of messages to return from each stream.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>A value of <see cref="RedisStream"/> for each stream.</returns>
-        /// <remarks>Equivalent of calling XREADGROUP GROUP groupName consumerName COUNT countPerStream STREAMS stream1 stream2 id1 id2</remarks>
-        /// <remarks>https://redis.io/commands/xreadgroup</remarks>
+        /// <remarks>
+        /// <para>Equivalent of calling <c>XREADGROUP GROUP groupName consumerName COUNT countPerStream STREAMS stream1 stream2 id1 id2</c>.</para>
+        /// <para><seealso href="https://redis.io/commands/xreadgroup"/></para>
+        /// </remarks>
         RedisStream[] StreamReadGroup(StreamPosition[] streamPositions, RedisValue groupName, RedisValue consumerName, int? countPerStream, CommandFlags flags);
 
         /// <summary>
@@ -2340,8 +2420,10 @@ namespace StackExchange.Redis
         /// <param name="noAck">When true, the message will not be added to the pending message list.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>A value of <see cref="RedisStream"/> for each stream.</returns>
-        /// <remarks>Equivalent of calling XREADGROUP GROUP groupName consumerName COUNT countPerStream STREAMS stream1 stream2 id1 id2</remarks>
-        /// <remarks>https://redis.io/commands/xreadgroup</remarks>
+        /// <remarks>
+        /// <para>Equivalent of calling <c>XREADGROUP GROUP groupName consumerName COUNT countPerStream STREAMS stream1 stream2 id1 id2</c>.</para>
+        /// <para><seealso href="https://redis.io/commands/xreadgroup"/></para>
+        /// </remarks>
         RedisStream[] StreamReadGroup(StreamPosition[] streamPositions, RedisValue groupName, RedisValue consumerName, int? countPerStream = null, bool noAck = false, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -2352,7 +2434,7 @@ namespace StackExchange.Redis
         /// <param name="useApproximateMaxLength">If true, the "~" argument is used to allow the stream to exceed max length by a small number. This improves performance when removing messages.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The number of messages removed from the stream.</returns>
-        /// <remarks>https://redis.io/topics/streams-intro</remarks>
+        /// <remarks><seealso href="https://redis.io/topics/streams-intro"/></remarks>
         long StreamTrim(RedisKey key, int maxLength, bool useApproximateMaxLength = false, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -2363,7 +2445,7 @@ namespace StackExchange.Redis
         /// <param name="value">The value to append to the string.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The length of the string after the append operation.</returns>
-        /// <remarks>https://redis.io/commands/append</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/append"/></remarks>
         long StringAppend(RedisKey key, RedisValue value, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -2377,7 +2459,7 @@ namespace StackExchange.Redis
         /// <param name="end">The end byte to count at.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The number of bits set to 1.</returns>
-        /// <remarks>https://redis.io/commands/bitcount</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/bitcount"/></remarks>
         long StringBitCount(RedisKey key, long start = 0, long end = -1, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -2392,7 +2474,7 @@ namespace StackExchange.Redis
         /// <param name="second">The second key to get the bit value from.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The size of the string stored in the destination key, that is equal to the size of the longest input string.</returns>
-        /// <remarks>https://redis.io/commands/bitop</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/bitop"/></remarks>
         long StringBitOperation(Bitwise operation, RedisKey destination, RedisKey first, RedisKey second = default, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -2405,7 +2487,7 @@ namespace StackExchange.Redis
         /// <param name="keys">The keys to get the bit values from.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The size of the string stored in the destination key, that is equal to the size of the longest input string.</returns>
-        /// <remarks>https://redis.io/commands/bitop</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/bitop"/></remarks>
         long StringBitOperation(Bitwise operation, RedisKey destination, RedisKey[] keys, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -2419,9 +2501,11 @@ namespace StackExchange.Redis
         /// <param name="start">The position to start looking (defaults to 0).</param>
         /// <param name="end">The position to stop looking (defaults to -1, unlimited).</param>
         /// <param name="flags">The flags to use for this operation.</param>
-        /// <returns>The command returns the position of the first bit set to 1 or 0 according to the request.
-        /// If we look for set bits(the bit argument is 1) and the string is empty or composed of just zero bytes, -1 is returned.</returns>
-        /// <remarks>https://redis.io/commands/bitpos</remarks>
+        /// <returns>
+        /// The command returns the position of the first bit set to 1 or 0 according to the request.
+        /// If we look for set bits(the bit argument is 1) and the string is empty or composed of just zero bytes, -1 is returned.
+        /// </returns>
+        /// <remarks><seealso href="https://redis.io/commands/bitpos"/></remarks>
         long StringBitPosition(RedisKey key, bool bit, long start = 0, long end = -1, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -2434,8 +2518,10 @@ namespace StackExchange.Redis
         /// <param name="value">The amount to decrement by (defaults to 1).</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The value of key after the decrement.</returns>
-        /// <remarks>https://redis.io/commands/decrby</remarks>
-        /// <remarks>https://redis.io/commands/decr</remarks>
+        /// <remarks>
+        /// <seealso href="https://redis.io/commands/decrby"/>,
+        /// <seealso href="https://redis.io/commands/decr"/>
+        /// </remarks>
         long StringDecrement(RedisKey key, long value = 1, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -2447,7 +2533,7 @@ namespace StackExchange.Redis
         /// <param name="value">The amount to decrement by (defaults to 1).</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The value of key after the decrement.</returns>
-        /// <remarks>https://redis.io/commands/incrbyfloat</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/incrbyfloat"/></remarks>
         double StringDecrement(RedisKey key, double value, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -2457,7 +2543,7 @@ namespace StackExchange.Redis
         /// <param name="key">The key of the string.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The value of key, or nil when key does not exist.</returns>
-        /// <remarks>https://redis.io/commands/get</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/get"/></remarks>
         RedisValue StringGet(RedisKey key, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -2467,7 +2553,7 @@ namespace StackExchange.Redis
         /// <param name="keys">The keys of the strings.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The values of the strings with nil for keys do not exist.</returns>
-        /// <remarks>https://redis.io/commands/mget</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/mget"/></remarks>
         RedisValue[] StringGet(RedisKey[] keys, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -2477,7 +2563,7 @@ namespace StackExchange.Redis
         /// <param name="key">The key of the string.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The value of key, or nil when key does not exist.</returns>
-        /// <remarks>https://redis.io/commands/get</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/get"/></remarks>
         Lease<byte>? StringGetLease(RedisKey key, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -2488,7 +2574,7 @@ namespace StackExchange.Redis
         /// <param name="offset">The offset in the string to get a bit at.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The bit value stored at offset.</returns>
-        /// <remarks>https://redis.io/commands/getbit</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/getbit"/></remarks>
         bool StringGetBit(RedisKey key, long offset, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -2501,7 +2587,7 @@ namespace StackExchange.Redis
         /// <param name="end">The end index of the substring to get.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The substring of the string value stored at key.</returns>
-        /// <remarks>https://redis.io/commands/getrange</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/getrange"/></remarks>
         RedisValue StringGetRange(RedisKey key, long start, long end, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -2511,7 +2597,7 @@ namespace StackExchange.Redis
         /// <param name="value">The value to replace the existing value with.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The old value stored at key, or nil when key did not exist.</returns>
-        /// <remarks>https://redis.io/commands/getset</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/getset"/></remarks>
         RedisValue StringGetSet(RedisKey key, RedisValue value, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -2522,7 +2608,7 @@ namespace StackExchange.Redis
         /// <param name="expiry">The expiry to set. <see langword="null"/> will remove expiry.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The value of key, or nil when key does not exist.</returns>
-        /// <remarks>https://redis.io/commands/getex</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/getex"/></remarks>
         RedisValue StringGetSetExpiry(RedisKey key, TimeSpan? expiry, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -2533,7 +2619,7 @@ namespace StackExchange.Redis
         /// <param name="expiry">The exact date and time to expire at. <see cref="DateTime.MaxValue"/> will remove expiry.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The value of key, or nil when key does not exist.</returns>
-        /// <remarks>https://redis.io/commands/getex</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/getex"/></remarks>
         RedisValue StringGetSetExpiry(RedisKey key, DateTime expiry, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -2544,7 +2630,7 @@ namespace StackExchange.Redis
         /// <param name="key">The key of the string.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The value of key, or nil when key does not exist.</returns>
-        /// <remarks>https://redis.io/commands/getdelete</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/getdelete"/></remarks>
         RedisValue StringGetDelete(RedisKey key, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -2555,7 +2641,7 @@ namespace StackExchange.Redis
         /// <param name="key">The key of the string.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The value of key and its expiry, or nil when key does not exist.</returns>
-        /// <remarks>https://redis.io/commands/get</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/get"/></remarks>
         RedisValueWithExpiry StringGetWithExpiry(RedisKey key, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -2568,8 +2654,10 @@ namespace StackExchange.Redis
         /// <param name="value">The amount to increment by (defaults to 1).</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The value of key after the increment.</returns>
-        /// <remarks>https://redis.io/commands/incrby</remarks>
-        /// <remarks>https://redis.io/commands/incr</remarks>
+        /// <remarks>
+        /// <seealso href="https://redis.io/commands/incrby"/>,
+        /// <seealso href="https://redis.io/commands/incr"/>
+        /// </remarks>
         long StringIncrement(RedisKey key, long value = 1, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -2581,7 +2669,7 @@ namespace StackExchange.Redis
         /// <param name="value">The amount to increment by (defaults to 1).</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The value of key after the increment.</returns>
-        /// <remarks>https://redis.io/commands/incrbyfloat</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/incrbyfloat"/></remarks>
         double StringIncrement(RedisKey key, double value, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -2590,7 +2678,7 @@ namespace StackExchange.Redis
         /// <param name="key">The key of the string.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The length of the string at key, or 0 when key does not exist.</returns>
-        /// <remarks>https://redis.io/commands/strlen</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/strlen"/></remarks>
         long StringLength(RedisKey key, CommandFlags flags = CommandFlags.None);
 
         /// <inheritdoc cref="StringSet(RedisKey, RedisValue, TimeSpan?, bool, When, CommandFlags)" />
@@ -2611,7 +2699,7 @@ namespace StackExchange.Redis
         /// <param name="when">Which condition to set the value under (defaults to always).</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns><see langword="true"/> if the string was set, <see langword="false"/> otherwise.</returns>
-        /// <remarks>https://redis.io/commands/set</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/set"/></remarks>
         bool StringSet(RedisKey key, RedisValue value, TimeSpan? expiry = null, bool keepTtl = false, When when = When.Always, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -2622,8 +2710,10 @@ namespace StackExchange.Redis
         /// <param name="when">Which condition to set the value under (defaults to always).</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns><see langword="true"/> if the keys were set, <see langword="false"/> otherwise.</returns>
-        /// <remarks>https://redis.io/commands/mset</remarks>
-        /// <remarks>https://redis.io/commands/msetnx</remarks>
+        /// <remarks>
+        /// <seealso href="https://redis.io/commands/mset"/>,
+        /// <seealso href="https://redis.io/commands/msetnx"/>
+        /// </remarks>
         bool StringSet(KeyValuePair<RedisKey, RedisValue>[] values, When when = When.Always, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -2635,8 +2725,10 @@ namespace StackExchange.Redis
         /// <param name="when">Which condition to set the value under (defaults to <see cref="When.Always"/>).</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The previous value stored at <paramref name="key"/>, or nil when key did not exist.</returns>
-        /// <remarks>This method uses the SET command with the GET option introduced in Redis 6.2.0 instead of the deprecated GETSET command.</remarks>
-        /// <remarks>https://redis.io/commands/set</remarks>
+        /// <remarks>
+        /// <para>This method uses the <c>SET</c> command with the <c>GET</c> option introduced in Redis 6.2.0 instead of the deprecated <c>GETSET</c> command.</para>
+        /// <para><seealso href="https://redis.io/commands/set"/></para>
+        /// </remarks>
         RedisValue StringSetAndGet(RedisKey key, RedisValue value, TimeSpan? expiry, When when, CommandFlags flags);
 
         /// <summary>
@@ -2650,7 +2742,7 @@ namespace StackExchange.Redis
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The previous value stored at <paramref name="key"/>, or nil when key did not exist.</returns>
         /// <remarks>This method uses the SET command with the GET option introduced in Redis 6.2.0 instead of the deprecated GETSET command.</remarks>
-        /// <remarks>https://redis.io/commands/set</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/set"/></remarks>
         RedisValue StringSetAndGet(RedisKey key, RedisValue value, TimeSpan? expiry = null, bool keepTtl = false, When when = When.Always, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -2663,7 +2755,7 @@ namespace StackExchange.Redis
         /// <param name="bit">The bit value to set, true for 1, false for 0.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The original bit value stored at offset.</returns>
-        /// <remarks>https://redis.io/commands/setbit</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/setbit"/></remarks>
         bool StringSetBit(RedisKey key, long offset, bool bit, CommandFlags flags = CommandFlags.None);
 
         /// <summary>
@@ -2676,7 +2768,7 @@ namespace StackExchange.Redis
         /// <param name="value">The value to overwrite with.</param>
         /// <param name="flags">The flags to use for this operation.</param>
         /// <returns>The length of the string after it was modified by the command.</returns>
-        /// <remarks>https://redis.io/commands/setrange</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/setrange"/></remarks>
         RedisValue StringSetRange(RedisKey key, long offset, RedisValue value, CommandFlags flags = CommandFlags.None);
     }
 }
