+++ src/StackExchange.Redis/Interfaces/IRedis.cs
@@ -12,7 +12,7 @@ namespace StackExchange.Redis
         /// </summary>
         /// <param name="flags">The command flags to use when pinging.</param>
         /// <returns>The observed latency.</returns>
-        /// <remarks>https://redis.io/commands/ping</remarks>
+        /// <remarks><seealso href="https://redis.io/commands/ping"/></remarks>
         TimeSpan Ping(CommandFlags flags = CommandFlags.None);
     }
 }
