+++ src/StackExchange.Redis/Interfaces/ITransaction.cs
@@ -5,14 +5,14 @@ namespace StackExchange.Redis
     /// <summary>
     /// Represents a group of operations that will be sent to the server as a single unit,
     /// and processed on the server as a single unit. Transactions can also include constraints
-    /// (implemented via WATCH), but note that constraint checking involves will (very briefly)
-    /// block the connection, since the transaction cannot be correctly committed (EXEC),
-    /// aborted (DISCARD) or not applied in the first place (UNWATCH) until the responses from
+    /// (implemented via <c>WATCH</c>), but note that constraint checking involves will (very briefly)
+    /// block the connection, since the transaction cannot be correctly committed (<c>EXEC</c>),
+    /// aborted (<c>DISCARD</c>) or not applied in the first place (<c>UNWATCH</c>) until the responses from
     /// the constraint checks have arrived.
     /// </summary>
-    /// <remarks>https://redis.io/topics/transactions</remarks>
     /// <remarks>
-    /// Note that on a cluster, it may be required that all keys involved in the transaction (including constraints) are in the same hash-slot.
+    /// <para>Note that on a cluster, it may be required that all keys involved in the transaction (including constraints) are in the same hash-slot.</para>
+    /// <para><seealso href="https://redis.io/topics/transactions"/></para>
     /// </remarks>
     public interface ITransaction : IBatch
     {
